+++ stripe/src/test/java/com/stripe/android/IssuingCardPinServiceTest.java
@@ -15,7 +15,6 @@ import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
 
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyMapOf;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -41,7 +40,7 @@ public class IssuingCardPinServiceTest {
             "}";
 
     @Mock
-    private StripeApiHandler.RequestExecutor mRequestExecutor;
+    private RequestExecutor mRequestExecutor;
     @Mock
     IssuingCardPinService.IssuingCardPinRetrievalListener mockRetrievalListener;
     @Mock
@@ -75,7 +74,7 @@ public class IssuingCardPinServiceTest {
                         null);
 
         when(mRequestExecutor.execute(
-                eq(StripeApiHandler.RestMethod.GET),
+                eq(RequestExecutor.RestMethod.GET),
                 eq("https://api.stripe.com/v1/issuing/cards/ic_abcdef/pin"),
                 ArgumentMatchers.<String, Object>anyMap(),
                 any(RequestOptions.class)
@@ -103,7 +102,7 @@ public class IssuingCardPinServiceTest {
                         null);
 
         when(mRequestExecutor.execute(
-                eq(StripeApiHandler.RestMethod.POST),
+                eq(RequestExecutor.RestMethod.POST),
                 eq("https://api.stripe.com/v1/issuing/cards/ic_abcdef/pin"),
                 ArgumentMatchers.<String, Object>anyMap(),
                 any(RequestOptions.class)
@@ -133,7 +132,7 @@ public class IssuingCardPinServiceTest {
                         null);
 
         when(mRequestExecutor.execute(
-                eq(StripeApiHandler.RestMethod.GET),
+                eq(RequestExecutor.RestMethod.GET),
                 eq("https://api.stripe.com/v1/issuing/cards/ic_abcdef/pin"),
                 ArgumentMatchers.<String, Object>anyMap(),
                 any(RequestOptions.class)
